Move more logic into TorchAgent.
